improved install & uninstall scripts
The installer will only run if it detects it has not run before.
The uninstaller will remove the .dotfiles config file on
uninstall so the installer knows its safe to run again.

+++ install.sh
@@ -1,24 +1,30 @@
 #!/bin/bash
 
-# Remove the .dotfiles config file if it exists already; someone is running
-# install multiple times.
-# Create the .dotfiles config file so we can use the path for our functions 
-rm $PWD/.dotfiles 2> /dev/null
-echo -e "# Record the path where we installed the dot files; needed for our functions\n${PWD}" >> .dotfiles
+# Do not run if the .dotfiles config file is present; already installed
+if [ -f $PWD/.dotfiles ]; then
+    
+    echo "Already installed, uninstall first."
+    
+else
 
-# A block list of files that should not be symbolically linked.
-declare -A block=([.functions]=1 [.gitignore]=1)
+    # Create the .dotfiles config file so we can use the path for our functions 
+    echo -e "# Record the path where we installed the dot files. This is needed for our\n# functions and prevents duplicate installation.\n${PWD}" >> .dotfiles
 
-# Find all dot files then if the original file exists, create a backup
-# Once backed up to {file}.dtbak symlink the new dotfile in place
-for file in $(find . -maxdepth 1 -name ".*" -type f  -printf "%f\n" ); do
-    # Do not create symbolic links for files in the block list
-    if [ ! ${block[$file]} ]; then
-        if [ -e ~/$file ]; then
-            mv -f ~/$file{,.dtbak}
+    # A block list of files that should not be symbolically linked
+    declare -A block=([.functions]=1 [.gitignore]=1)
+
+    # Find all dot files then if the original file exists, create a backup
+    # Once backed up to {file}.dtbak symlink the new dotfile in place
+    for file in $(find . -maxdepth 1 -name ".*" -type f  -printf "%f\n" ); do
+        # Do not create symbolic links for files in the block list
+        if [ ! ${block[$file]} ]; then
+            if [ -e ~/$file ]; then
+                mv -f ~/$file{,.dtbak}
+            fi
+            ln -s $PWD/$file ~/$file
         fi
-        ln -s $PWD/$file ~/$file
-    fi
-done
+    done
 
-echo "Installed"
+    echo "Installed"
+
+fi
